test-kafka/pyproject.toml
[project]
name = "test-kafka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hsiang-Jen Li", email = "hsiangjenli@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.3.0",
    "ollama>=0.6.1",
]

[project.scripts]
test-kafka = "test_kafka:main"

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
